adjust class
