+++ flutter/beagle/lib/beagle_server_driven_state.dart
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+class BeagleServerDrivenState {
+  BeagleServerDrivenState._();
+
+  factory BeagleServerDrivenState.started() = BeagleServerDrivenStateStarted;
+
+  factory BeagleServerDrivenState.success() = BeagleServerDrivenStateSuccess;
+
+  factory BeagleServerDrivenState.finished() = BeagleServerDrivenStateFinished;
+
+  //NOT IMPLEMENTED
+  factory BeagleServerDrivenState.canceled() = BeagleServerDrivenStateCanceled;
+
+  factory BeagleServerDrivenState.error(Exception exception) =
+      BeagleServerDrivenStateError;
+}
+
+class BeagleServerDrivenStateSuccess extends BeagleServerDrivenState {
+  BeagleServerDrivenStateSuccess() : super._();
+}
+
+class BeagleServerDrivenStateStarted extends BeagleServerDrivenState {
+  BeagleServerDrivenStateStarted() : super._();
+}
+
+class BeagleServerDrivenStateFinished extends BeagleServerDrivenState {
+  BeagleServerDrivenStateFinished() : super._();
+}
+
+class BeagleServerDrivenStateCanceled extends BeagleServerDrivenState {
+  BeagleServerDrivenStateCanceled() : super._();
+}
+
+class BeagleServerDrivenStateError extends BeagleServerDrivenState {
+  BeagleServerDrivenStateError(this.exception) : super._();
+
+  final Exception exception;
+}
